+++ shared-components/index.ts
@@ -1,1 +1,1 @@
-export { default as DateField } from "./DateField";
+export { default as DateField } from "./Form/DateField";
